Small Changes & Brazier
[+] Removed the enemies from the scene
[+] Add enemies
[+] Moved rhythm bar

+++ Assets/_SCRIPTS_/Enterprise Video/EnemyMoveToPlayer.cs
@@ -22,7 +22,7 @@ public class EnemyMoveToPlayer : MonoBehaviour
         {
             Move();
 
-            timer -= MusicRhythmTimer.SecondsBetweenBeats();
+            timer -= (float)MusicRhythmTimer.SecondsBetweenBeats();
         }
     }
